+++ src/Process/ProcessException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Solital\Core\Process;
+
+class ProcessException extends \Exception
+{
+}